+++ backend/src/routes/controller/post.courses.js
@@ -1,0 +1,39 @@
+
+const { Course,Teacher } = require("../../db.js");
+const { getCategoryId } = require('./getCategoryId.js')
+
+const postCourses = async (req,res) => {
+
+    //*email is of the teacher, category must be an array
+    const { name, description, email, img, price, category } = req.body;
+    try {
+      const FK = await Teacher.findOne({
+        //Busca el id del teacher
+        where: {
+          email: email,
+        },
+      });
+      const courseCreated = await Course.create({
+        //Crea el curso
+        name,
+        description,
+        price,
+        img,
+        FKteacherID: FK.id,
+        score: "5",
+      });
+      const categoryID = await getCategoryId(category); //Busca el id de las categorias
+      // console.log('category id in post course:',categoryID);
+      courseCreated.addCategory(categoryID); //Agrega las categorias al curso
+      res.status(200).send(courseCreated);
+    } catch (error) {
+      console.error(error);
+      res.status(404).send(error);
+    }
+  };
+
+
+
+module.exports = {
+  postCourses,
+}
