+++ backend/src/routes/index.js
@@ -1,5 +1,4 @@
 const { Router } = require("express");
-
 // Importar todos los routers;
 const register = require("./register.js");
 const login = require("./localLogin.js");
@@ -10,25 +9,25 @@ const review = require("./review.js");
 const video = require("./video.js");
 const buyCourse = require("./buyCourse.js");
 const home = require("./home.js");
+const students = require("./students.js");
+const teachers = require("./teachers.js");
 
 const router = Router();
 
 const user = require("./user.routes");
 
 // Configurar los routers
-router.use("/register", register);
-router.use("/fakecourses", fakeCourses);
-router.use("/courses", courses);
-router.use("/category", category);
-router.use("/review", review);
-router.use("/video", video);
 router.use("/buy", buyCourse);
-
-//In this route the front will send the id of the student and the back will give back the last five videos the user has watched
-router.use("/home", home);
-
-
+router.use("/fakecourses", fakeCourses);
+router.use("/category", category);
+router.use("/courses", courses);
+router.use("/home", home);//In this route the front will send the id of the student and the back will give back the last five videos the user has watched
 router.use("/login", login);
+router.use("/register", register);
+router.use("/review", review);
+router.use("/students", students);
+router.use("/teachers", teachers);
+router.use("/video", video);
 
 
 /////////////////USER////////////////
