+++ backend/src/routes/students.js
@@ -1,0 +1,9 @@
+const { Router } = require("express");
+
+const router = Router();
+
+const {getStudents} = require('./controller/students/getStudents')
+
+router.get('/', getStudents); 
+
+module.exports = router;
